optimize command flow

+++ src/api/lite_renderer.c
@@ -75,11 +75,11 @@ static int f_draw_rect(lua_State* L)
 
 static int f_draw_text(lua_State* L)
 {
-    LiteFont**   font   = luaL_checkudata(L, 1, API_TYPE_FONT);
+    LiteFont**  font   = luaL_checkudata(L, 1, API_TYPE_FONT);
     const char* text   = luaL_checkstring(L, 2);
-    int         x      = luaL_checknumber(L, 3);
-    int         y      = luaL_checknumber(L, 4);
-    LiteColor    color  = checkcolor(L, 5, 255);
+    int         x      = (int)luaL_checknumber(L, 3);
+    int         y      = (int)luaL_checknumber(L, 4);
+    LiteColor   color  = checkcolor(L, 5, 255);
     int         next_x = lite_rencache_draw_text(*font, text, x, y, color);
     lua_pushnumber(L, next_x);
     return 1;
