+++ src/api/lite_renderer_font.c
@@ -5,7 +5,7 @@
 static int f_load(lua_State* L)
 {
     const char* filename = luaL_checkstring(L, 1);
-    float       size     = luaL_checknumber(L, 2);
+    float       size     = (float)luaL_checknumber(L, 2);
     LiteFont**  self     = lua_newuserdata(L, sizeof(*self));
     luaL_setmetatable(L, API_TYPE_FONT);
     *self = lite_load_font(filename, size);
